requirements.txt
Django==1.8
argparse==1.2.1
django-extensions==1.6.1
django-tenant-schemas==1.5.8
djangorestframework==3.3.2
psycopg2==2.6.1
six==1.10.0
wsgiref==0.1.2
